Wrote some comments and refactored reorder_projection
