+++ sql/analyzer/optimization_rules.go
@@ -51,141 +51,6 @@ func optimizeDistinct(ctx *sql.Context, a *Analyzer, node sql.Node, scope *Scope
 	return node, nil
 }
 
-func reorderProjection(ctx *sql.Context, a *Analyzer, n sql.Node, scope *Scope) (sql.Node, error) {
-	span, ctx := ctx.Span("reorder_projection")
-	defer span.Finish()
-
-	if n.Resolved() {
-		return n, nil
-	}
-
-	return plan.TransformUp(n, func(node sql.Node) (sql.Node, error) {
-		project, ok := node.(*plan.Project)
-		// When we transform the projection, the children will always be
-		// unresolved in the case we want to fix, as the reorder happens just
-		// so some columns can be resolved.
-		// For that, we need to account for NaturalJoin, whose schema can't be
-		// obtained until it's resolved and ignore the projection for the
-		// moment until the resolve_natural_joins has finished resolving the
-		// node and we can tackle it in the next iteration.
-		// Without this check, it would cause a panic, because NaturalJoin's
-		// schema method is just a placeholder that should not be called.
-		if !ok || hasNaturalJoin(project.Child) {
-			return node, nil
-		}
-
-		// We must find all columns that may need to be moved inside the
-		// projection.
-		var newColumns = make(map[string]sql.Expression)
-		for _, col := range project.Projections {
-			alias, ok := col.(*expression.Alias)
-			if ok {
-				newColumns[alias.Name()] = col
-			}
-		}
-
-		// And add projection nodes where needed in the child tree.
-		var didNeedReorder bool
-		child, err := plan.TransformUp(project.Child, func(node sql.Node) (sql.Node, error) {
-			var requiredColumns []string
-			switch node := node.(type) {
-			case *plan.Sort, *plan.Filter:
-				for _, expr := range node.(sql.Expressioner).Expressions() {
-					sql.Inspect(expr, func(e sql.Expression) bool {
-						if e != nil && e.Resolved() {
-							return true
-						}
-
-						uc, ok := e.(column)
-						if ok && uc.Table() == "" {
-							if _, ok := newColumns[uc.Name()]; ok {
-								requiredColumns = append(requiredColumns, uc.Name())
-							}
-						}
-
-						return true
-					})
-				}
-			default:
-				return node, nil
-			}
-
-			if len(requiredColumns) == 0 {
-				return node, nil
-			}
-
-			didNeedReorder = true
-
-			// Only add the required columns for that node in the projection.
-			child := node.Children()[0]
-			schema := child.Schema()
-			var projections = make([]sql.Expression, 0, len(schema)+len(requiredColumns))
-			for i, col := range schema {
-				projections = append(projections, expression.NewGetFieldWithTable(
-					i, col.Type, col.Source, col.Name, col.Nullable,
-				))
-			}
-
-			for _, col := range requiredColumns {
-				if c, ok := newColumns[col]; ok {
-					projections = append(projections, c)
-					delete(newColumns, col)
-				}
-			}
-
-			child = plan.NewProject(projections, child)
-			switch node := node.(type) {
-			case *plan.Filter:
-				return plan.NewFilter(node.Expression, child), nil
-			case *plan.Sort:
-				return plan.NewSort(node.SortFields, child), nil
-			default:
-				return nil, ErrInvalidNodeType.New("reorderProjection", node)
-			}
-		})
-
-		if err != nil {
-			return nil, err
-		}
-
-		if !didNeedReorder {
-			return project, nil
-		}
-
-		child, err = resolveColumns(ctx, a, child, scope)
-		if err != nil {
-			return nil, err
-		}
-
-		childSchema := child.Schema()
-		// Finally, replace the columns we moved with GetFields since they
-		// have already been projected.
-		var projections = make([]sql.Expression, len(project.Projections))
-		for i, p := range project.Projections {
-			if alias, ok := p.(*expression.Alias); ok {
-				var found bool
-				for idx, col := range childSchema {
-					if col.Name == alias.Name() {
-						projections[i] = expression.NewGetField(
-							idx, col.Type, col.Name, col.Nullable,
-						)
-						found = true
-						break
-					}
-				}
-
-				if !found {
-					projections[i] = p
-				}
-			} else {
-				projections[i] = p
-			}
-		}
-
-		return plan.NewProject(projections, child), nil
-	})
-}
-
 func moveJoinConditionsToFilter(ctx *sql.Context, a *Analyzer, n sql.Node, scope *Scope) (sql.Node, error) {
 	if !n.Resolved() {
 		return n, nil
@@ -422,18 +287,4 @@ func isTrue(e sql.Expression) bool {
 		}
 	}
 	return false
-}
-
-// hasNaturalJoin checks whether there is a natural join at some point in the
-// given node and its children.
-func hasNaturalJoin(node sql.Node) bool {
-	var found bool
-	plan.Inspect(node, func(node sql.Node) bool {
-		if _, ok := node.(*plan.NaturalJoin); ok {
-			found = true
-			return false
-		}
-		return true
-	})
-	return found
 }
