+++ sql/analyzer/optimization_rules_test.go
@@ -11,129 +11,6 @@ import (
 	"github.com/liquidata-inc/go-mysql-server/sql/plan"
 )
 
-func TestReorderProjection(t *testing.T) {
-	f := getRule("reorder_projection")
-
-	table := memory.NewTable("mytable", sql.Schema{{
-		Name: "i", Source: "mytable", Type: sql.Int64,
-	}})
-
-	testCases := []struct {
-		name     string
-		project  sql.Node
-		expected sql.Node
-	}{
-		{
-			"sort",
-			plan.NewProject(
-				[]sql.Expression{
-					expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-					expression.NewAlias("foo", expression.NewLiteral(1, sql.Int64)),
-					expression.NewAlias("bar", expression.NewLiteral(2, sql.Int64)),
-				},
-				plan.NewSort(
-					[]plan.SortField{
-						{Column: expression.NewUnresolvedColumn("foo")},
-					},
-					plan.NewFilter(
-						expression.NewEquals(
-							expression.NewLiteral(1, sql.Int64),
-							expression.NewUnresolvedColumn("bar"),
-						),
-						plan.NewResolvedTable(table),
-					),
-				),
-			),
-			plan.NewProject(
-				[]sql.Expression{
-					expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-					expression.NewGetField(2, sql.Int64, "foo", false),
-					expression.NewGetField(1, sql.Int64, "bar", false),
-				},
-				plan.NewSort(
-					[]plan.SortField{{Column: expression.NewGetField(2, sql.Int64, "foo", false)}},
-					plan.NewProject(
-						[]sql.Expression{
-							expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-							expression.NewGetField(1, sql.Int64, "bar", false),
-							expression.NewAlias("foo", expression.NewLiteral(1, sql.Int64)),
-						},
-						plan.NewFilter(
-							expression.NewEquals(
-								expression.NewLiteral(1, sql.Int64),
-								expression.NewGetField(1, sql.Int64, "bar", false),
-							),
-							plan.NewProject(
-								[]sql.Expression{
-									expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-									expression.NewAlias("bar", expression.NewLiteral(2, sql.Int64)),
-								},
-								plan.NewResolvedTable(table),
-							),
-						),
-					),
-				),
-			),
-		},
-		{
-			"use alias twice",
-			plan.NewProject(
-				[]sql.Expression{
-					expression.NewAlias("foo", expression.NewLiteral(1, sql.Int64)),
-				},
-				plan.NewFilter(
-					expression.NewOr(
-						expression.NewEquals(
-							expression.NewLiteral(1, sql.Int64),
-							expression.NewUnresolvedColumn("foo"),
-						),
-						expression.NewEquals(
-							expression.NewLiteral(1, sql.Int64),
-							expression.NewUnresolvedColumn("foo"),
-						),
-					),
-					plan.NewResolvedTable(table),
-				),
-			),
-			plan.NewProject(
-				[]sql.Expression{
-					expression.NewGetField(1, sql.Int64, "foo", false),
-				},
-				plan.NewFilter(
-					expression.NewOr(
-						expression.NewEquals(
-							expression.NewLiteral(1, sql.Int64),
-							expression.NewGetField(1, sql.Int64, "foo", false),
-						),
-						expression.NewEquals(
-							expression.NewLiteral(1, sql.Int64),
-							expression.NewGetField(1, sql.Int64, "foo", false),
-						),
-					),
-					plan.NewProject(
-						[]sql.Expression{
-							expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-							expression.NewAlias("foo", expression.NewLiteral(1, sql.Int64)),
-						},
-						plan.NewResolvedTable(table),
-					),
-				),
-			),
-		},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.name, func(t *testing.T) {
-			require := require.New(t)
-
-			result, err := f.Apply(sql.NewEmptyContext(), NewDefault(nil), tt.project, nil)
-			require.NoError(err)
-
-			require.Equal(tt.expected, result)
-		})
-	}
-}
-
 func TestEraseProjection(t *testing.T) {
 	require := require.New(t)
 	f := getRule("erase_projection")
